win32 int -> intptr_t changes for 64 bit warning

+++ Tools/QuickLink/src/QL_Interface.cpp
@@ -337,7 +337,7 @@ void QL_Interface::spawn_debugger()
 
 	// spawn the debugger asynchronously
 	
-	int ret = _spawnlp(_P_NOWAIT, "javaw.exe", "javaw.exe", "-jar", "SoarJavaDebugger.jar", "-remote", NULL);
+	intptr_t ret = _spawnlp(_P_NOWAIT, "javaw.exe", "javaw.exe", "-jar", "SoarJavaDebugger.jar", "-remote", NULL);
 	if(ret == -1) {
 		switch (errno) {
 				case E2BIG:
